v1.0.2: lean tool profile, honest results, Dynamo cold start
Token usage
- The server registers everything but enables a 14-tool core profile
  (~3,000 tokens instead of ~32,000 for the full catalogue). The new
  revit_tools tool lists and enables groups (query, arch, mep, views,
  annotate, modify, dynamo, memory) and the server sends
  notifications/tools/list_changed. REVIT_MCP_PROFILE=core|standard|full,
  REVIT_MCP_TOOLS adds groups at start-up.
- Every tool description and schema rewritten short; shared schema
  fragments in server/src/utils/schemas.ts; conventions in
  server/TOOL-CONVENTIONS.md.
- Compact results through server/src/utils/reply.ts: no indentation,
  nulls/empty strings/UniqueId pruned, floats rounded to 0.001 mm, listing
  tools default to 30 records with a total, oversized arrays cut to whole
  records and marked _truncated (REVIT_MCP_MAX_RESULT_CHARS, default 20000).
- Server "instructions" (units, never invent ids, how to enable tools, what
  an error looks like, undo semantics, what to do when Revit is unreachable)
  are handed to the client at initialise time.

Fixed
- Six commands answered with the previous call's result because they never
  reset their completion signal: dynamo_op, delete_element,
  get_selected_elements, get_current_view_info, get_available_family_types,
  say_hello.
- get_available_family_types, get_current_view_info and get_selected_elements
  kept a previous error forever; the plugin's "background thread" diagnosis
  was attached to every error instead of Revit API exceptions only.
- Every failure now reaches the client as isError with {"ok": false, ...},
  including Revit replies of the shape {success: false} / {Success: false};
  before, 5 of 101 tools did.
- delete_element accepts numeric ids and 64-bit ids (Revit 2024+).
- ai_element_filter bounding box schema matches the handler (single point,
  not a two-point line), so spatial filters no longer match nothing.
- Results carry metric values with the unit in the key (areaM2, volumeM3,
  perimeterMm, unboundedHeightMm, totalAreaM2, totalVolumeM3, LocationMm,
  StartMm, EndMm, LengthMm) where internal feet had been returned under a
  millimetre promise.

Audit repair (AUDIT-2026-09-03.md)
- All 86 Revit commands audited tool-by-tool against their C# command,
  handler and model; 182 findings confirmed by adversarial verification, 184
  fixes across 70 commands: schema keys the handler never read
  (create_reference_plane, create_group, create_column, create_railing,
  create_model_curve, create_stair, create_roof, create_opening), success
  reported with nothing done (about 40 handlers), parameters silently
  ignored, null dereferences before the null check, unit and enum
  mismatches, descriptions that promised what the handler does not do.
- Pre-release review on top of the sweep: thirteen more batch creators
  (create_text_note, create_column, create_surface_based_element,
  create_conduit, create_direct_shape, create_equipment, create_mep_system,
  create_pipe, create_space, create_swept_shape, create_schedule,
  create_sheet, create_view) still answered Success=true with an empty
  list; they now fail with the skip reasons, and six others no longer say
  "Successfully created 0" on a failure. A tool in no catalogue group was
  never enabled even under REVIT_MCP_PROFILE=full (latent; fixed).
  transform_elements' description matches what the handler returns.
  query_geometry, query_references, query_view_range and
  analyze_model_statistics now carry the unit in the key too (MinMm/MaxMm,
  AreaM2, SurfaceAreaM2, VolumeM3, LengthMm, OffsetMm, elevationMm), and
  create_mep_curve passes the nearest level to Revit instead of an invalid
  level id. dynamo_run_graph reports an evaluation that did not finish
  within timeout_seconds as an error instead of ok. manage_project_parameters
  list reports the real group label and visibility on Revit 2023+ (an
  unreachable #elif had hardcoded PG_DATA / true). export_views no longer
  lists each PNG/JPG once per view of the batch; create_tag warns (and
  fails when nothing was tagged) on a null from IndependentTag.Create;
  create_detail_curve reports a partial batch as success with the shortfall
  and its description no longer restricts it to drafting/plan views;
  create_dimensions options no longer drop integer values (JSON integers
  arrive as long); create_group no longer reports an empty batch as success.

Dynamo
- dynamo_run_graph honours timeout_seconds (sent as timeoutMs), refuses to
  run when a different workspace is open, launches Dynamo itself when it is
  not running (new dynamo_op launch, posts ID_VISUAL_PROGRAMMING_DYNAMO),
  polls the new eval_status op and returns every node that did not end
  Active; a Warning/Error node makes the result an error.
- A closed Dynamo window (live model object, disposed engine) is reported as
  not running instead of failing every open with a NullReferenceException.
- dynamo_edit_graph: create: true starts from an empty Dynamo 3 graph;
  add_function_node places library nodes by DesignScript signature with
  named ports.

Add-in
- The socket service starts at ApplicationInitialized; settings.autoStart
  false or REVIT_MCP_AUTOSTART=0 restores the manual switch.

Tooling and release
- tools/Deploy-Local.ps1 refreshes the AppData server copy and the Revit
  Addins folder; build/utils/selfTest.js checks reply helpers and catalogue
  and Verify.ps1 runs it; Make-Installer.ps1 splats a hashtable (it never
  built a Setup.exe before); Inno Setup 7 is found; the Set-RevitMcpTarget
  self-test runs on PowerShell 5.1; .gitattributes fixes line endings.
- release.yml also attaches mcp-servers-for-revit-<ver>-server.zip; the
  npm-publish job is disabled on this fork (the npm name belongs upstream).

Fork identity and install path
- READ ME FIRST.txt (generated by Package.ps1), the installer's finish page,
  Start-menu link and support URLs, the add-in manifest's vendor link,
  server/package.json and both READMEs point at MrGezz/mcp-servers-for-revit
  and describe this fork's install: bundled server registered by path,
  socket that starts with Revit, lean tool profile. The upstream project
  stays credited; "npx -y mcp-server-for-revit" (upstream's 1.0.0 on npm)
  is no longer presented as the way to run this add-in's server.
- Set-RevitMcpTarget.ps1 -IncludeClaudeCode wrote to ~/.claude/settings.json,
  which Claude Code does not read for MCP servers (user scope is
  ~/.claude.json); it now runs "claude mcp add --scope user". Measured in a
  sandboxed CLAUDE_CONFIG_DIR on PowerShell 5.1.
- Node.js floor is 20 everywhere (package.json engines already said >=20;
  README, installer check and CI said 18).
- RevitMCPPlugin.dll no longer names the upstream author as Company with a
  2025 copyright; RevitMCPCommandSet.dll carries the release version instead
  of 1.0.0.0; the Setup.exe publisher is MrGezz. scripts/release.ps1 tags the
  current branch (it used to check out and hard-reset main) and bumps the
  command set's <Version>.
- README: the registry is seeded with every command enabled on first start,
  so Settings is optional; Revit 2027 / .NET 10 listed; release steps name
  features/icz-addin; every Supported Tools row shows its revit_tools group.

Verified: tools/Verify.ps1 ALL PASS (26 checks: years agree in 7 places,
payload shape for 8 Revit versions, tsc, Dynamo harness, reply/catalog
harness, PowerShell 5.1 parse and ASCII, Set-RevitMcpTarget self-test, ISCC
compile). Live against Revit 2026: say_hello, an unknown category returns a
clean error and the next call succeeds, dynamo_status.

+++ commandset/Services/DeleteElementEventHandler.cs
@@ -49,7 +49,10 @@ namespace RevitMCPCommandSet.Services
 
                 foreach (var idStr in ElementIds)
                 {
-                    if (int.TryParse(idStr, out int elementIdValue))
+                    // long, not int: Revit 2024+ ids exceed Int32 in large models, and an
+                    // id that fails to parse here was reported as "unparseable" and never
+                    // deleted while the caller had sent a perfectly valid number.
+                    if (long.TryParse(idStr, out long elementIdValue))
                     {
                         var elementId = ElementIdFactory.Create(elementIdValue);
                         if (doc.GetElement(elementId) != null)
